Experimental readonly support
When run with -readonly mode, database is open in read-only mode. This
is still exclusive access to file, you cannot have another process
writing to it, but can have multiple processes all open the same
database in read-only mode.

This change is for a very specific use case, it only disables writing to
on-disk file, it does not disable data-modification commands, it's just
all changes would stay in memory and won't be purged to disk. Over time
such changes may increase memory usage.