Match Redis return value for BITOP AND command with empty result
Redis bitop commands return [1] “the size of the string stored in the
destination key is equal to the size of the longest input string.”

Tests elsewhere may depend on this exact semantics, so make an attempt to
preserve it, even though it doesn't affect the actual result (created bitmap).

Before this change, when one of the keys was empty (non-existing), the resulting
bitmap was empty and the return value of this operation was 0 (the “empty string”
length). With this change, the effect of the operation stays the same (storing
and empty bitmap), but as the return value we return “the size of the longest
input string” as Redis does.

[1]: https://redis.io/docs/latest/commands/bitop/#return-information