Fix BITOP AND handling of mixing existing and non-existing keys
Follow-up for e29eefc66a756ccf21da4a96b4ddc6d915710d33, make sure that order
of existing vs non-existing keys in BITOP AND operation doesn't matter. While
here, simplify tracking of which “bitmap string” are considered to be empty and
which aren't that defines whether the destination bitmap is stored or not.